+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/service/ApplicationService.kt
@@ -4,15 +4,9 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.ApplicationSubmitted
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.ApplicationSubmittedEnvelope
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.ApplicationSubmittedSubmittedBy
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.ApplicationWithdrawn
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.ApplicationWithdrawnEnvelope
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.Ldu
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.PersonReference
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.Region
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.events.model.Team
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ApplicationSortField
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ApplicationTimelineNote
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.Cas1ApplicationUserDetails
@@ -24,7 +18,6 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.TimelineEvent
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.WithdrawalReason
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.client.ApDeliusContextApiClient
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.client.ClientResult
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.client.CommunityApiClient
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.config.NotifyConfig
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.ApAreaRepository
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.ApplicationEntity
@@ -51,14 +44,13 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.PersonInfoResult
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.PersonRisks
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.ValidationErrors
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.community.OffenderDetailSummary
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.community.StaffUserDetails
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.deliuscontext.CaseDetail
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.prisonsapi.InOutStatus
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.model.validated
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.results.AuthorisableActionResult
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.results.CasResult
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.results.ValidatableActionResult
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.WithdrawableState
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.cas1.Cas1ApplicationDomainEventService
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.transformer.ApplicationTimelineNoteTransformer
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.transformer.ApplicationTimelineTransformer
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.transformer.AssessmentClarificationNoteTransformer
@@ -69,7 +61,6 @@ import java.time.Instant
 import java.time.LocalDate
 import java.time.LocalTime
 import java.time.OffsetDateTime
-import java.time.Period
 import java.time.ZoneOffset
 import java.util.UUID
 import javax.transaction.Transactional
@@ -95,7 +86,6 @@ class ApplicationService(
   private val applicationTimelineNoteTransformer: ApplicationTimelineNoteTransformer,
   private val domainEventService: DomainEventService,
   private val cas3DomainEventService: Cas3DomainEventService,
-  private val communityApiClient: CommunityApiClient,
   private val apDeliusContextApiClient: ApDeliusContextApiClient,
   private val applicationTeamCodeRepository: ApplicationTeamCodeRepository,
   private val emailNotificationService: EmailNotificationService,
@@ -108,6 +98,7 @@ class ApplicationService(
   private val applicationTimelineTransformer: ApplicationTimelineTransformer,
   private val domainEventTransformer: DomainEventTransformer,
   private val cas1ApplicationUserDetailsRepository: Cas1ApplicationUserDetailsRepository,
+  private val cas1ApplicationDomainEventService: Cas1ApplicationDomainEventService,
 ) {
   fun getApplication(applicationId: UUID) = applicationRepository.findByIdOrNull(applicationId)
 
@@ -813,7 +804,7 @@ class ApplicationService(
     assessmentService.createApprovedPremisesAssessment(application)
     application = applicationRepository.save(application)
 
-    createApplicationSubmittedEvent(application, submitApplication, username, jwt)
+    cas1ApplicationDomainEventService.applicationSubmitted(application, submitApplication, username, jwt)
     if (user.email != null) {
       sendEmailApplicationSubmitted(user, application)
     }
@@ -917,143 +908,6 @@ class ApplicationService(
 
     return AuthorisableActionResult.Success(
       ValidatableActionResult.Success(application),
-    )
-  }
-
-  private fun createApplicationSubmittedEvent(
-    application: ApprovedPremisesApplicationEntity,
-    submitApplication: SubmitApprovedPremisesApplication,
-    username: String,
-    jwt: String,
-  ) {
-    val domainEventId = UUID.randomUUID()
-    val eventOccurredAt = OffsetDateTime.now()
-
-    val offenderDetails =
-      when (val offenderDetailsResult = offenderService.getOffenderByCrn(application.crn, username, true)) {
-        is AuthorisableActionResult.Success -> offenderDetailsResult.entity
-        is AuthorisableActionResult.Unauthorised ->
-          throw RuntimeException(
-            "Unable to get Offender Details when creating Application" +
-              "Submitted Domain Event: Unauthorised",
-          )
-
-        is AuthorisableActionResult.NotFound ->
-          throw RuntimeException(
-            "Unable to get Offender Details when creating Application" +
-              " Submitted Domain Event: Not Found",
-          )
-      }
-
-    val risks =
-      when (val riskResult = offenderService.getRiskByCrn(application.crn, jwt, username)) {
-        is AuthorisableActionResult.Success -> riskResult.entity
-        is AuthorisableActionResult.Unauthorised ->
-          throw RuntimeException("Unable to get Risks when creating Application Submitted Domain Event: Unauthorised")
-
-        is AuthorisableActionResult.NotFound ->
-          throw RuntimeException("Unable to get Risks when creating Application Submitted Domain Event: Not Found")
-      }
-
-    val mappaLevel = risks.mappa.value?.level
-
-    val staffDetails = when (val staffDetailsResult = communityApiClient.getStaffUserDetails(username)) {
-      is ClientResult.Success -> staffDetailsResult.body
-      is ClientResult.Failure -> staffDetailsResult.throwException()
-    }
-
-    val caseDetail = when (val caseDetailResult = apDeliusContextApiClient.getCaseDetail(application.crn)) {
-      is ClientResult.Success -> caseDetailResult.body
-      is ClientResult.Failure -> caseDetailResult.throwException()
-    }
-
-    domainEventService.saveApplicationSubmittedDomainEvent(
-      DomainEvent(
-        id = domainEventId,
-        applicationId = application.id,
-        crn = application.crn,
-        occurredAt = eventOccurredAt.toInstant(),
-        data = ApplicationSubmittedEnvelope(
-          id = domainEventId,
-          timestamp = eventOccurredAt.toInstant(),
-          eventType = "approved-premises.application.submitted",
-          eventDetails = getApplicationSubmittedForDomainEvent(
-            application,
-            offenderDetails,
-            mappaLevel,
-            submitApplication,
-            staffDetails,
-            caseDetail,
-          ),
-        ),
-      ),
-    )
-  }
-
-  private fun getApplicationSubmittedForDomainEvent(
-    application: ApprovedPremisesApplicationEntity,
-    offenderDetails: OffenderDetailSummary,
-    mappaLevel: String?,
-    submitApplication: SubmitApprovedPremisesApplication,
-    staffDetails: StaffUserDetails,
-    caseDetail: CaseDetail,
-  ): ApplicationSubmitted {
-    return ApplicationSubmitted(
-      applicationId = application.id,
-      applicationUrl = applicationUrlTemplate
-        .replace("#id", application.id.toString()),
-      personReference = PersonReference(
-        crn = application.crn,
-        noms = offenderDetails.otherIds.nomsNumber ?: "Unknown NOMS Number",
-      ),
-      deliusEventNumber = application.eventNumber,
-      mappa = mappaLevel,
-      offenceId = application.offenceId,
-      releaseType = submitApplication.releaseType.toString(),
-      age = Period.between(offenderDetails.dateOfBirth, LocalDate.now()).years,
-      gender = when (offenderDetails.gender.lowercase()) {
-        "male" -> ApplicationSubmitted.Gender.male
-        "female" -> ApplicationSubmitted.Gender.female
-        else -> throw RuntimeException("Unknown gender: ${offenderDetails.gender}")
-      },
-      targetLocation = submitApplication.targetLocation,
-      submittedAt = Instant.now(),
-      submittedBy = getApplicationSubmittedSubmittedBy(staffDetails, caseDetail),
-      sentenceLengthInMonths = null,
-    )
-  }
-
-  private fun getApplicationSubmittedSubmittedBy(
-    staffDetails: StaffUserDetails,
-    caseDetail: CaseDetail,
-  ): ApplicationSubmittedSubmittedBy {
-    return ApplicationSubmittedSubmittedBy(
-      staffMember = domainEventTransformer.toStaffMember(staffDetails),
-      probationArea = domainEventTransformer.toProbationArea(staffDetails),
-      team = getTeamFromCaseDetail(caseDetail),
-      ldu = getLduFromCaseDetail(caseDetail),
-      region = getRegionFromStaffDetails(staffDetails),
-    )
-  }
-
-  private fun getLduFromCaseDetail(caseDetail: CaseDetail): Ldu {
-    return Ldu(
-      code = caseDetail.case.manager.team.ldu.code,
-      name = caseDetail.case.manager.team.ldu.name,
-    )
-  }
-
-  private fun getTeamFromCaseDetail(caseDetail: CaseDetail): Team {
-    return Team(
-      code = caseDetail.case.manager.team.code,
-      name = caseDetail.case.manager.team.name,
-    )
-  }
-
-  private fun getRegionFromStaffDetails(staffDetails: StaffUserDetails): Region {
-    return Region(
-      code = staffDetails.probationArea.code,
-      name = staffDetails.probationArea.description,
     )
   }
